Added support for configuring the XJC Ant task backing the plugin.
The JAXB tools (e.g. XJC) bundled with the JDK are relocated to a package not
matching the JAXB-RI. There are a lot of XJC plugins around compiled against
the JAXB-RI which cannot be used with the JAXB tools bundled with the JDK due
to this. When configuring the plugin to use the JAXB-RI Ant task instead of the
Ant task bundled with the JDK (e.g. com.sun.tools.xjc.XJC2Task), those plugins
can be used.